Create Sidebar.tsx
