+++ src/components/Sidebar/AppLogo.scss
@@ -1,0 +1,3 @@
+#appLogo {
+  background-color: #fafafa;
+}
