+++ src/components/Sidebar/AppLogo.tsx
@@ -1,0 +1,12 @@
+import "./AppLogo.scss"
+
+export default function AppLogo(){
+  return (
+    <div id="appLogo" className="flex items-center justify-center h-64 w-64">
+      <div className="text-center">
+        <h1 className="text-2xl">Typle</h1>
+        <h2 className="text-xs mt-4">IMPROVE YOUR TYPING SKILLS</h2>
+      </div>
+    </div>
+  )
+}
