+++ src/components/Sidebar/NavListItem.scss
@@ -1,0 +1,8 @@
+li {
+  a {
+    &:hover {
+      background: #FEF0F6;
+      color: #E9176F;
+    }
+  }
+}
