+++ src/components/Sidebar/Sidebar.scss
@@ -1,0 +1,3 @@
+#sidebar {
+  background-color: #fff;
+}
